included the correlation in the sentiment.ipynb file
Reduced news data size to include only companies present in the financial data files. Performed sentiment analysis on the relevant news headlines. Merged financial data for 7 companies and converted the date to a pandas datetime format, separating date and time. Merged the news and financial data based on date alignment. Calculated correlation and visualized the results with a heatmap.